Update patch set 1
Patch Set 1:

> Are we using our own build of jiri yet?  Feels a bit dodgy to be
 > doing major surgery before we're actually using the thing and
 > making sure it does what we need it to do, etc.

We are not.  We have to fix some bugs related to bootstrapping before we can do that.

WRT testing, jiri has a relatively broad test suite, which passes for this change.  I also tested using it to sync the fuchsia manifest.  The bigger risk IMO is not having automated testing for jiri (since we moved it out of the vanadium repo).

Patch-set: 1
Label: Code-Review=0