Update patch set 4
Patch Set 4:

> The Go that's "special" is only the one that's on Googlers machines
 > from what I can tell (unless upstream debian/ubuntu strips this
 > package in a similar way).  Pretty sure most folks have one that
 > works fine.

We do control the go package that the bots will have, and that will be sane.

By "special", I was referring to the "and one we control" part of your comment.  I'm happy to tell folks that they need a go package from golang.org/dl in order to run this script.

I guess the upshot of this is I'm changing the default to "darwin linux" and if it fails, I'll tell the dev to download a go package capable of cross-compiling (and point them to golang.org/dl)

Patch-set: 4
Label: Code-Review=0