Update patch set 1
Patch Set 1:

We've never had the pre-".jiri_root" structure in fuchsia so adding this dir is not useful to anyone.  This should free up the top-level 'devtools' directory for whatever we want to use.

Chris - what tests for jiri are expected to pass right now and what setup should I use to run them in?

Patch-set: 1