Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> So this does the same as the "Pull" option in Gerrit's "Download" tab?  What state does that leave your branch in?

Yes, pretty much except that you can also use just the change ID (e.g. jiri cl patch 8367) and it'll find out what the latest CL is and patch it into your project (i.e. the same as what git cl does). You also don't have to be in the project the patch was sent to, Jiri can figure which project it should apply it to.

Right now, it'll create a new branch that's named after the "Pull" ref just by striping the "refs/" prefix, e.g. for 8367 it'll create a branch named changes/67/8367/2 (since the patchset is part of the name, you can use this to pull in different patchsets). If you have a better suggestion, I'm all ears. One thing I was considering would be to also strip the "changes/[0-9]+" part and leave just "8367/2".

If the branch already exists, it'll do nothing and give you an error message. You can pass in the -delete flag in which case it'll try to delete it first. If there are unmerged changes on that branch, you'll get an error as well. You can also add -force flag to get around that (basically it's a difference between using git branch -d and -D).

Patch-set: 2