Update patch set 2
Patch Set 2:

I liked the '-b' parameter to 'git cl' that let me pick the name.

This behavior seems like a fine default, but it needs to be documented as it's not predictable from the name of the command itself and it's very difficult to figure out what happened to a git repo after-the-fact.  Can you make 'jiri cl help patch' list out what it does?

Patch-set: 2
Label: Code-Review=0