Fix TO-50 "jiri runp -show-*-prefix=true" fails
It was failing as above command sets "-collate-stdout" to true which
tries to make a tmp file using the name of the project. In fuchsia we
have nested projects so this file creation fails.

To fix this we now create a tmp file which doesn't contain project
name.

I have also added a new line after runp output of each project for
improving user readability

Change-Id: I46f70cf615dc05ae37d76bdacfc597a71c111fd6