Fix TO-50 "jiri runp -show-*-prefix=true" fails
It was failing as above command sets "-collate-stdout" to true which
tries to make a tmp file using the name of directory. In fuchsia we have
nested projects so this file creation fails.
To fix this we now create a tmp directory first and then replicate our
folder structure in it.
I have also added a new line after runp output of each project for
improving user exp

Change-Id: I46f70cf615dc05ae37d76bdacfc597a71c111fd6