Update patch set 1
Patch Set 1:

> I was thinking about adding an option to jiri update that would
 > force it use the local manifest repository rather than pulling the
 > upstream copy (which could include your changes), that should make
 > testing local changes much easier.

I talked to nlacasse about this at some point and he seemed to think that was non-trivial.  Though that may be because of some workflow that we no longer support.  I definitely want that change, but in the meantime, there are folks asking how to test local changes and AFAIK this is the only way.

Patch-set: 1
Label: Code-Review=0