Update patch set 4
Patch Set 4:

The bug in TO-46 seems to be due to jiri using the latest snapshot to speed up the project scan.  This is fine, except in cases where the user has mangled the local projects somehow and the snapshot no longer matches in a way that jiri can't handle.

Patch-set: 4