Update patch set 1
Abandoned

there's really no good way to get around the installed location being root-owned since linux mv semantics depend on the directory permissions.  we could replace the renames with copies, but then we run the risk of screwing up the install if the user kills the process.

the official solution is to not install jiri in a root-owned directory

Patch-set: 1
Status: abandoned