Update patch set 2
Patch Set 2:

This is IMHO the wrong approach, we shouldn't be hacking around the command line argument parsing (especially since we'll be eventually replacing that part), we should be instead just calling runImport with the right arguments. We might also consider breaking it into a few functions a testing those separately.

Patch-set: 2