Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> It is not hacking per se, I am just resetting the commands object to their original state. This will also help us test that code uses correct default values and also how it will behave under global variables. I can change it to what you are suggesting...please advise...

Commands were not really designed to be used in this way (and the replacement I'm working on is going to break this completely). I'd really prefer to avoid testing Commands and flag parsing and just focus on the logic itself. That's what we're already doing in tests for all other commands, runp and import are the only exceptions and this is merely just about making them consistent with the rest.

Patch-set: 2
Label: Code-Review=0