Make project info actually useful
This makes several changes to how project info command. First of
all, it provides sane default output (other than the name). The
regexp name matching is now optional same as the template, and the
template is executed against well defined type, not an internal one.
Finally there is now also an option to emit JSON output which can be
used for machine processing (e.g. on bots).

Change-Id: Ib463c894bdfd55eedac98907b185a15cbccf579a