Adding new upload command
Writing jiri upload command from scratch. This version will just act as a
wrapper on git commands and simply push current changes to gerrit. We
plan to make it multi-part CLs in future changes.
When it is ready we will swap it with current jiri cl.

To run this try jiri upload

Change-Id: I85f8e533e2b85a9e06e1998b6e82d8253327df8c