Update patch set 2
Patch Set 2:

RE: "TO-76: Jiri shouldn't say "not rebasing untracked branch" if that branch is
up-to-date"

After some thought, it would still be nice to get a NOTE message for every untracked branch you have. But say it's up to date if it is; otherwise suggest manual rebase instructions as you do already.

Patch-set: 2
Label: Code-Review=0