Update patch set 2
Patch Set 2:

> RE: "TO-76: Jiri shouldn't say "not rebasing untracked branch" if
 > that branch is
 > up-to-date"
 > 
 > After some thought, it would still be nice to get a NOTE message
 > for every untracked branch you have. But say it's up to date if it
 > is; otherwise suggest manual rebase instructions as you do already.

But that will cause lot of noise for people who work on multiple projects. And that will also hide the rebase messages which are actually important. Like for me I have around 20-25 branches in various project, as I don't delete them, that will cause lot of noise

Patch-set: 2