Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > PTAL
> 
> The problem with your change is that you're conflating two different Git concepts which is the exact opposite of what we want to do, that is making Jiri more transparent.
> 
> Your change treats the following as two distinct options:
> 
> git clone --reference path/to/ref
> git clone --shared path/to/repo
> 
> However, the following is perfectly valid option in Git as well:
> 
> git clone --reference path/to/ref --shared path/to/repo

To be more clear, Git doesn't really concept of a cache, so that itself is already a misnomer, but I think it's fine to use it for reference repositories since that's the common use case. However, treating a shared repository as a "type of cache" is wrong because reference and shared options are two separate and orthogonal concepts.

Patch-set: 11
Reviewer: Gerrit User 5050 <5050@ccdefb6c-23ea-33ff-848c-339fc6ca2269>
Label: Code-Review=0