Write a .git/JIRI_HEAD pointing each project's upstream ref.
As a user it can be hard to know what ref jiri thinks is the right one
to be synced to. Often origin/master isn't right because the manifest
has a project tracking a different branch or pinned to a specific
commit. This change makes it so that a user can do commands like `git
checkout JIRI_HEAD` or `git rebase JIRI_HEAD`.

Change-Id: Ia269544bbafc8b18412af29b1e416541806f0df9