Update patch set 4
Patch Set 4:

(1 comment)

I've looked at this and decided that since I know nothing about Go, I don't have the ability to review it. The tests are written in a way that looks suspicious to me, with a lot of loops and constructing error messages instead of using an assertion library, but according to my research that kind of thing is considered good style in Go, so I can't comment on it.

Patch-set: 4
Reviewer: Gerrit User 9007 <9007@ccdefb6c-23ea-33ff-848c-339fc6ca2269>