Correct import tool documentation wrt flags.
If you try to use "-path" as a command-line flag while importing,
jiri complains that there's no such flag. Jiri understands "root",
however, and that seems to have the desired effect, so let's have
the docs reflect this.

(If someone more familiar with the code wants to make "path" a
valid argument, that'd be dandy--I poked around a bit, but it seems
like "path" is created only when the project is *loaded*, and it
seem there to be created ad-hoc via a ProjectKey--at that point I
wasn't sure whether there was some important semantic distinction
between project *root* and project *path* and opted to just change
the docs to something that works.)

Change-Id: Ic24a00874d729bf3ab4d205e564f3ed72023e8c9