Don't leave behind tmp dirs when running snapshot
Before loader used to return the temp directory it created and the
caller had to delete that directory. Moving that logic to jiri context
so that all the cleanup can happen at one place, as not all loader
functions return temp directory.

Also updating warning message to tell user to check their manifest files

TO-635

Change-Id: I3181d2e6a3f3ff74ba19a5d30a483f8e4e66b7c6