Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 2:
> > > > 
> > > > Why not make it an argument as in the case of `update`?
> > > 
> > > It already accepts an argument which is the output file name.
> > 
> > IMO I'd make the output file name an argument (e.g. -json-output for consistency with other comamnds?) and then use the argument for snapshot. We'll probably need to do it in two steps: first add the option that has the same behavior as the argument, update recipe, and then change the argument behavior. WDYT?
> 
> SGTM, so add both -json-output AND -snapshot?

What I was suggesting is to first make another change that adds -json-output flag which would be an alias for the argument. We'll roll it out and update the recipe to start using -json-output when producing source manifest. Then in this change we can change the semantics of the argument to load snapshot.

Patch-set: 6
Reviewer: Gerrit User 5050 <5050@ccdefb6c-23ea-33ff-848c-339fc6ca2269>