Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 4:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > > Patch Set 2:
> > > > > 
> > > > > Why not make it an argument as in the case of `update`?
> > > > 
> > > > It already accepts an argument which is the output file name.
> > > 
> > > IMO I'd make the output file name an argument (e.g. -json-output for consistency with other comamnds?) and then use the argument for snapshot. We'll probably need to do it in two steps: first add the option that has the same behavior as the argument, update recipe, and then change the argument behavior. WDYT?
> > 
> > SGTM, so add both -json-output AND -snapshot?
> 
> What I was suggesting is to first make another change that adds -json-output flag which would be an alias for the argument. We'll roll it out and update the recipe to start using -json-output when producing source manifest. Then in this change we can change the semantics of the argument to load snapshot.

I mentioned this earlier but there is already precedent for using the argument as the output file: the `jiri snapshot` command and so in the face of that I'm not convinced -json-output is going to clean much up here? jiri update, however, accepts the snapshot as an argument, so IDK.

I just wanted to address that since the goal for this is to have a more consistent interface to Jiri.

Patch-set: 6
Reviewer: Gerrit User 5210 <5210@ccdefb6c-23ea-33ff-848c-339fc6ca2269>