Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > > Patch Set 2:
> > > > > 
> > > > > Why not make it an argument as in the case of `update`?
> > > > 
> > > > It already accepts an argument which is the output file name.
> > > 
> > > Just kidding, this doesn't actually work yet because NewSourceManifest expects there to be a full checkout. I don't really understand why, considering that the Projects structure already tells you everything you need to know?
> > 
> > OK turns out fetch_ref is optional so I modified NewSourceManifest to only grab it when necessary, also revision.
> 
> it is optional, but someone was using it, so i had to add that dependency

does this mean we shouldn't ignore a failure to try to pick it up?

I need functionality that looks something like this (gives me a way to inspect and iterate over a snapshot-like thing in JSON because recipes) for releases, so if this isn't the way to go do you have any suggestions?

Patch-set: 6
Reviewer: Gerrit User 5210 <5210@ccdefb6c-23ea-33ff-848c-339fc6ca2269>