Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> What's the motivation for this? I'm slightly concerned about the increase in the amount of data we'll need to pull with this change.

As far as I understand, on all the bots, the first time jiri is run, it creates a cache, which is basically git clone --mirror
From then on for any future operation, it will first check that cache to see if it has the necessary revisions
The basic problem is that we're telling jiri to set --depth when creating the cache, and then later operations try to refer to commits that are farther back in the history than --depth

So to fix this the plan is to do 2 things:
1. Have jiri set --no-single-branch when it creates the cache. This will make it pull in all the branches into the cache, not just master. When used with --depth, it will respect --depth for every branch separately.
2. Have our infra automatically create branches for every commit going forward.

This effectively will force jiri to cache all commits that happen from the point we start step 2
And this will all be unnecessary once IN-621 is done. I can add a TODO to revert this change once that is done.

Patch-set: 3
CC: Gerrit User 18706 <18706@ccdefb6c-23ea-33ff-848c-339fc6ca2269>
Reviewer: Gerrit User 19891 <19891@ccdefb6c-23ea-33ff-848c-339fc6ca2269>
Label: Commit-Queue=0