Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Petr, could you please review this soon?
> 
> I'm fine with the change from technical point of view, my concern is that this is going to significantly increase the size of the checkout for repositories that use depth like WebKit, Pixman, or Cairo if we do this blindly. Have you done any measurements to checks what's the size increase?

Using the topaz manifest and running:
$ jiri init -cache="$PWD/git-cache" -shared -analytics-opt=false
$ jiri update
$ du -h --max-depth=0 git-cache
Before:
7.4G    git-cache
After:
8.6G    git-cache

So a 16% increase in checkout size. Not ideal but it doesn't seem prohibitive to me.

We could roll this change back after IN-621 is done. I can add a TODO to undo it in the code if you want.

Patch-set: 3
Reviewer: Gerrit User 19891 <19891@ccdefb6c-23ea-33ff-848c-339fc6ca2269>
Label: Commit-Queue=0