Add shallow_since option for jiri projects.
This is intended to eventually replace the historydepth option. Once
all manifests are switched over I will remove the historydepth code.

Background:
We currently use historydepth on a few repos that we've
forked that have lots of code and many years of history. We use it
to reduce the amount of time and resources that it takes to clone them,
especially when creating a mirror clone, which happens when creating a
local cache.

Unfortunately this breaks when some job runs on a bot that has the cache
and refers to a commit that is older than the historydepth. This has
happened a few times already, see INTK-485.

Using shallow-since instead of history allows us to set this flag to
some date and pick up all commits after that date and not those commits
before that date. If we set the value to something old enough
(e.g. 1 year), then it's extremely unlikely that we'll ever have a
build that fails in the above scenario, but we also won't be pulling in
gigabytes of unneeded git history when creating the cache.

Bug: IN-621 #comment
Change-Id: I75b448d35335b50624585de29d4fe8f71e31e2e2