Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> (1 comment)

I am not going to merge it unless I have a way to test it on GI bots. Though this patch should not break existing projects as long as there is no package declaration in manifest files, which is not going to happen before full cipd support is implemented and tested. 

I am working on adding a cipd bootstrap function to jiri so it can fetch the latest cipd binary when it is needed.

Patch-set: 8
Reviewer: Gerrit User 12756 <12756@ccdefb6c-23ea-33ff-848c-339fc6ca2269>
Label: Commit-Queue=0