[cipd] add platforms attribute for packages
This patch adds platforms attribute for <package> tags in manifests.
Currently jiri does not generate locks for linux-arm64 packages
as not all packages are targeted for linux-arm64. The newly added
'platforms' allows user to specify the exact platforms the packages
are targeted, which make it possible for jiri to generate correct
locks for additional platforms including linux-arm64.

Change-Id: Icb0bc5c615c329f10f688f55a1c209e92105cbf1