[snapshort] Using package instanceID as version when checkout a snapshot
This patch changes jiri's behavior when checking out a snapshot file.
Before this patch, jiri use the version tag of a package when checking
out a snapshot file, which is causing some issues as the tag of a cipd
packages can be changed. After this patch, jiri will use instance ids
directly for packages when checking out a snapshot file when these ids
are available. This patch also changed the behavior of creating
a snapshot file to always use local manifests.

Test: Local

Change-Id: I2834c54b8f362d69e8c6f35fc633078d30c8af16