[lockfile] Add 'version' to lockfile.
This change adds 'version' tag to lockfile for packages. We discovered
that CIPD will not report mismatches of instance-id and package version
since CIPD will not use remote backend for verification if a lockfile
is used. Therefore, we added 'version' tag to allow jiri to report
any mismatches of the manifests and lockfiles.

Test: Local
Change-Id: I3549341a37ea944b2d97477f413658ec3f7221ba