Update patch set 1
Patch Set 1:

Hm, is this a requested feature? I am personally happy for jiri to tell me that there are uncommitted changes and to go do whatever myself (which is often to stash the changes rather than to commit anything). In cases where this is on an in-progress branch, this output could be pages long, and I often already know that's the case. So often for me, this change would be more annoying than helpful. But I don't know what others' use-cases for this are.

Perhaps we can gate this feature behind a flag?

Patch-set: 1
Reviewer: Gerrit User 30867 <30867@ccdefb6c-23ea-33ff-848c-339fc6ca2269>