[manifest] Fill default field values for packages
`jiri manifest` doesn't include non-empty default values for package
elements, which seems to have been a simple oversight in forgetting to
call FillDefaults() for the packages.

CIPD package rollers need to know the default platforms for a package so
that they know which CIPD packages to check when searching for a common
tag to pin all the packages to.

Bug: 60085
Change-Id: I9ea319cb8d91553094faa15933eb6a63aaf198a8