[partial] Introduce SkipPartial flag.
This will allow Infra to specify repositories that should not use
partial clone behavior at init time. This works around an issue with
partial clones where a Gerrit change has its checkout time increase with
the number of commits behind HEAD it is. We will therefore just not use
partial clones for any repo that a change will be patched into.

Bug: 34965
Change-Id: I226b77db7e7ebd0037c7261889a148b4952e53a9