[test] Add workaround for git default branch name change
Git implemented a default branch name change and it will break
existing jiri unit tests and potentially cause breakages in the
future if a new repository is added to the manifest. This change
adds a workaround to clear unit test until we have a better fix
for the this problem.

Bug: 65012

Change-Id: Id55a6f57d3839fca205767c753e499e3ac573283