[cipd] Include local path in package lock key
jiri should allow the same package to be included by a manifest multiple
times at different paths. However, this was previously not always
possible because jiri only used the name and version of each package to
compute it unique key when checking for duplicate entries in a lockfile.
So it was only possible to have the same package at two different paths
as long as the two were pinned to different versions.

That bug caused failures like this when a package was pinned to the same
version at two different paths:
https://ci.chromium.org/b/8841199368556918912

By including the package's local path in the key, we'll no longer
consider that to be a collision.

Change-Id: I3dde72473f43d6dbcd5cc3565f79e98c88694b64