[ops] Modify jiri project move behavior
Previously, `jiri` would attempt to move a directory atomically with
os.Rename(), falling back to exec'ing `mv` directly if the move crossed
a mount point.

This behavior does not handle a move of a directory to a child of
itself, in either the primary or fallback path. Instead we opt to route
all moves of directories thru a temporary swap directory in `.jiri_root`
to ensure we do not fall into any child path pitfalls and to avoid the
mount point complexities if we were to swap thru tmp.

Bug: b/195346750
Change-Id: Id99e3127c942f08e4da93d6a5ecf21189b3c8de5