[ops] Make the op order more permissive to moves
This ensures that parents are updated before children move up the
directory structure and after children are moved down the directory
structure to avoid cases where there are shifting boundries between the
parent/child.

This also makes the move logic more permissive and allows moving into an
existing directory if the repository is the only contents.

e.g third_party/foo/src/... will now be able to move to third_party/foo/...
so long as the src/ folder is the only contents of foo/

Change-Id: I88c4e5d7a077ddc13e221f5008df395b506fd748