[manifest] Update `jiri help manifest-files` output
Follow-up to fxr/688624.  I was under the impression that the output for
the jiri help manifest-files command was sourced from manifest.md, which
is not the case, it's sourced from strings defined in cmd.go and doc.go.
This will update those references as well to include documentation for
the attributes attribute.

Bug: 102021
Change-Id: I6e595c852e146e95d1eb72e778abcf918f4f6fa7