[update] Allow jiri updates to dissociate from the cache
Infrastructure occasionally hits failures when the git cache (global to
all builds run on a bot) is purged, but a checkout cached in the
incremental cache contains a reference to an object in said git cache.
This CL adds a CLI flag called `--dissociate` that is passed through to
the underlying git invocation when running a clone. As described by the
git documentation in https://git-scm.com/docs/git-clone, this option
uses the git cache to fetch objects (thus reducing network traffic), but
makes copies of borrowed objects. This allows us to drop the git cache
without leaving any dangling references in any of the incremental build
caches.

Bug: b/204319656
Change-Id: I172c12d7c390972bb77817f8d9e830373bb78ee2