[update] Allow jiri to dissociate from the cache part 2
This CL finishes the implementation of cache dissociation after an
update. See part 1 for a more detailed description for why this needs to
happen: https://fxr/c/jiri/+/785841.

As far as how we accomplish this, we run a dissociation manually on
fuchsia.git by running a repack and then removing the alternatives file.
This is how `git clone` implements the `--dissociate` flag (see the
implementation in
https://github.com/git/git/blob/master/builtin/clone.c#L1399), but we
can't use that since we run `git fetch` for fuchsia.git, which does not
support this flag.

Bug: b/204319656
Change-Id: I128fcfdd54a33c73ead1f5e27035fbdac075d08b