Ignore directories with .jiri-ignore file when traversing
For local work, it may be useful to include an unmanaged git repo in a
jiri repository. If that repo has a git submodule, "jiri update" can
fail with an error because the .git/jiri/metadata.v2 of the submodule
doesn't exist.

To prevent this from causing problems, support ignoring directories
containing a .jiri-ignore file.

Change-Id: I51247e397d03e657d9f671a886bc94c72aaac0d4