[jiri] Add `jiri init -exclude-dirs=`
This adds a <repeatable> init flag that takes a directory relative to
the root of the fuchsia checkout (top-level only) and excludes it from
local project searches. By default this uses "out" and "prebuilt" for
exclusions and should decouple `jiri update` from the number and size of
the user's build directories to some degree.

Fixed: b/287331288
Change-Id: Ibfb967dd89c1ea044b71829082606c22ebfba1ea