[ops] Swap update/create priority
Project updates should always happen before creations. The previous
logic worked based on path ordering in fuchsia.git because the jiri root
is fuchsia.git and all other projects are subprojects. However tests
often use sibling projects and the jiri root is not also a jiri project.
Whenever a create project op was the 0th element and its parent op was
not the 1st element, this would reault in incorrect sort ordering.

Change-Id: Ib465d61e546b5753b1f59cf7e291af1dd0a08fa3