[projects] Fix usage of LoadUpdatedManifest()
We need to pass a non-empty list of projects to this function, so we
must always unconditionally load the local project state, what is
returned from the function represents the full set of remote projects so
we much filter them based on configured attributes before attempting to
load project state in `jiri project`.

Also fix `-use-local-manifest`, we call the flag `-local-manifest` for
other subcommands, local consistenct trumps stylistic correctness.

Change-Id: I0d5b224229378bdde1e0cdd37a55477df7ef3681