Remove support for running arbitrary "jiri-*" scripts
Previously, if you ran `jiri foo` and had an executable named `jiri-foo`
on your $PATH, jiri would run that script. This was presumably intended
to mimic Git, which supports the same thing for `git foo` -> `git-foo`,
but it was an undocumented feature and I'm not aware of any people using
it, and can't find any "jiri-*" scripts in Code Search.

Change-Id: Ib30015c76ba6f021fa10411a2c1c188efe23322a