Introduce context.Context
Most of Jiri's code dates back to before the "context" library was
introduced in go1.7. In order to replace other custom parts of Jiri with
libraries, e.g. Jiri's custom subcommand library with
github.com/google/subcommands, we need to use contexts since those
libraries expect them.

Also this should eventually let us get rid of the weird `env` and
`jirix` stuff, which is not standard, in favor of contexts.

Change-Id: I2232f08e2c1dd546f914130f4766fac18434c854