[jiri] Move all flag values into structs
This greatly reduces the number of global variables and eases the
transition toward using a more standard subcommands library.

The worst part about the previous state was that the `package` command
didn't actually declare its own flags, and instead just reused flag
variables declared for the `project` command, which was an especially
easy mistake to make when there was no namespacing of flag variables.

Change-Id: I7745493c32c6314a8a5c44b34c0e3a9b855c6972