Migrate "update" command to subcommands library
The github.com/google/subcommands library provides essentially the same
functionality as Jiri's custom "cmdline" library while not storing state
in global variables, making it possible to run unit tests in parallel.

For now the `cmdline` library will continue to drive all the commands,
and we'll do a switch flip to fully migrate to the upstream library.

Bug: 356134056
Change-Id: I70be9d6510c912e7c9fe1ced9225b1d5b7d4990f