Fix `jiri package -json-output`
In commit 20fc8004e311df5977eab6f3a50477516f9068d1 I made the value of
the `-json-output` flag no longer be shared between `jiri package` and
`jiri project`. However, I missed that the two commands use the same
function for writing a JSON output file, and that function specifically
referenes the value of the `-json-output` flag from `jiri project`.

I updated the function to take in a path to write to instead of always
writing to the path referenced by the `project` command's `-json-output`
flag.

This had broken jiri usage in infra, which was fortunately caught in
presubmit by the recipe-tools-roller on https://fxrev.dev/1089585.

Change-Id: Iae507cff3b7e0608db71f22df68583a4446a5a36