Stop using os.Chdir()
`os.Chdir` is problematic because it mutates global state (the current
working directory), so code that uses it cannot safely run in parallel,
even in tests.

Replace direct `os.Chdir()` calls in production code with accesses of a
`jiri.X.Cwd` field that can be overridden in tests without affecting
global state, so that tests are better able to run in parallel.

Also migrate some tests away from changing the working directory
entirely.

Bug: 356134056
Change-Id: I8968ff2bdbb2fe0a3a9ae70edf890eb73fc99cf5