Disable `jiri init -cache` test
Jiri's cache feature is currently broken – running `jiri init
-cache=$(mktemp -d)` puts checkouts in a bad state such that `jiri
update` always fails.

For some reason this test was still passing in infra, but it's failing
locally, so disable it until we decide what to do about the cache
feature. It's not used in infra and I'm pretty sure it's never used
locally either, so there's a good chance we can (and should) delete it.

Change-Id: Ia0dc444f4b000e8ec29701f5e64a2412e755336f